rename class to PayoneHttpClientUtil
