+++ service/src/test/java/com/commercetools/util/PayoneHttpClientUtilTimeoutTest.java
@@ -19,13 +19,13 @@ import java.util.Random;
 
 import static com.commercetools.pspadapter.payone.domain.payone.PayonePostServiceImpl.executeGetRequest;
 import static com.commercetools.pspadapter.payone.domain.payone.PayonePostServiceImpl.executePostRequest;
-import static com.commercetools.util.PayoneHttpClientConfigurationUtil.*;
+import static com.commercetools.util.PayoneHttpClientUtil.*;
 import static java.lang.String.format;
 import static java.lang.Thread.sleep;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIOException;
 
-public class PayoneHttpClientConfigurationUtilTimeoutTest {
+public class PayoneHttpClientUtilTimeoutTest {
     /**
      * Url pattern to local mocking http service.
      * <ol>
@@ -35,7 +35,7 @@ public class PayoneHttpClientConfigurationUtilTimeoutTest {
      */
     private static final String URL_PATTERN = "http://localhost:%d?delay=%d";
 
-    private static Logger LOGGER = LoggerFactory.getLogger(PayoneHttpClientConfigurationUtilTimeoutTest.class);
+    private static Logger LOGGER = LoggerFactory.getLogger(PayoneHttpClientUtilTimeoutTest.class);
 
     private int testRandomPort;
 
@@ -111,7 +111,7 @@ public class PayoneHttpClientConfigurationUtilTimeoutTest {
     /**
      * Verify normal GET response.
      * This request has response timeout shorter than max request time in
-     * {@link PayoneHttpClientConfigurationUtil#TIMEOUT_TO_ESTABLISH_CONNECTION} thus should success, opposite to
+     * {@link PayoneHttpClientUtil#TIMEOUT_TO_ESTABLISH_CONNECTION} thus should success, opposite to
      * {@link #longGetRequest_shouldFailLong()}
      */
     @Test
@@ -124,9 +124,9 @@ public class PayoneHttpClientConfigurationUtilTimeoutTest {
 
     /**
      * Verify that if request timeout is too long - the client will retry 5 times
-     * ({@link PayoneHttpClientConfigurationUtil#RETRY_TIMES})
+     * ({@link PayoneHttpClientUtil#RETRY_TIMES})
      * This request has response timeout larger than max request time in
-     * {@link PayoneHttpClientConfigurationUtil#TIMEOUT_TO_ESTABLISH_CONNECTION} thus should failure, opposite to
+     * {@link PayoneHttpClientUtil#TIMEOUT_TO_ESTABLISH_CONNECTION} thus should failure, opposite to
      * {@link #shortGetRequest_shouldSuccess()}
      */
     @Test
@@ -144,10 +144,10 @@ public class PayoneHttpClientConfigurationUtilTimeoutTest {
     /**
      * Verify that if POST request timeout is too long - the client will fails with {@link java.io.IOException}, but
      * won't re-try, because the client is not re-sending request arguments once they are sent.
-     * See {@link PayoneHttpClientConfigurationUtil#REQUEST_SENT_RETRY_ENABLED} for more details.
+     * See {@link PayoneHttpClientUtil#REQUEST_SENT_RETRY_ENABLED} for more details.
      * <p>
      * This request has response timeout larger than max request time in
-     * {@link PayoneHttpClientConfigurationUtil#TIMEOUT_TO_ESTABLISH_CONNECTION} thus should failure, opposite to
+     * {@link PayoneHttpClientUtil#TIMEOUT_TO_ESTABLISH_CONNECTION} thus should failure, opposite to
      * {@link #shortGetRequest_shouldSuccess()}
      * <p>
      * This test should be change if we decide the request should be re-send on timeout exceptions.
